fix: 추상화 기반 다형성 구현
