+++ src/main/java/com/example/mini_project/domain/game/common/service/RankingRedisService.java
@@ -1,0 +1,12 @@
+package com.example.mini_project.domain.game.common.service;
+
+import com.example.mini_project.domain.game.common.dto.RankingResponseDto;
+import com.example.mini_project.domain.user.entity.User;
+
+import java.util.List;
+
+public interface RankingRedisService {
+    void createRanking(User user, int level, int gameScore);
+
+    List<RankingResponseDto> getTopRankings(int topN);
+}
